Combined Footer component and moved up.
