+++ src/applications/check-in/components/pages/demographics/DemographicsDisplay.jsx
@@ -8,6 +8,7 @@ export default function DemographicsDisplay({
   yesAction = () => {},
   noAction = () => {},
   Footer,
+  isPreCheckIn = true,
 }) {
   const demographicFields = [
     { title: 'Mailing address', key: 'mailingAddress' },
@@ -27,6 +28,7 @@ export default function DemographicsDisplay({
         yesAction={yesAction}
         noAction={noAction}
         Footer={Footer}
+        isPreCheckIn={isPreCheckIn}
       />
     </>
   );
