+++ src/applications/check-in/components/pages/emergencyContact/EmergencyContactDisplay.jsx
@@ -7,6 +7,7 @@ export default function EmergencyContactDisplay({
   noAction = () => {},
   isLoading,
   Footer,
+  isPreCheckIn = true,
 }) {
   const dataFields = [
     { title: 'Name', key: 'name' },
@@ -25,6 +26,7 @@ export default function EmergencyContactDisplay({
         noAction={noAction}
         Footer={Footer}
         isLoading={isLoading}
+        isPreCheckIn={isPreCheckIn}
       />
     </>
   );
