+++ src/applications/check-in/components/pages/nextOfKin/NextOfKinDisplay.jsx
@@ -9,6 +9,7 @@ export default function NextOfKinDisplay({
   noAction = () => {},
   isSendingData = false,
   Footer,
+  isPreCheckIn = true,
 }) {
   const nextOfKinFields = [
     { title: 'Name', key: 'name' },
@@ -38,6 +39,7 @@ export default function NextOfKinDisplay({
           );
         }}
         Footer={Footer}
+        isPreCheckIn={isPreCheckIn}
       />
     </>
   );
