--- src/applications/check-in/day-of/components/tests/Footer.unit.spec.jsx
@@ -1,71 +1,0 @@
-import React from 'react';
-import { expect } from 'chai';
-import { render } from '@testing-library/react';
-import Footer from '../Footer';
-
-describe('check-in', () => {
-  describe('Footer', () => {
-    it('Renders the footer with default props', () => {
-      const fakeStore = {
-        getState: () => ({
-          checkInData: {
-            appointments: [{ clinicPhoneNumber: '555-867-5309' }],
-          },
-        }),
-        subscribe: () => {},
-        dispatch: () => ({}),
-      };
-      const screen = render(<Footer store={fakeStore} />);
-
-      expect(screen.getByTestId('heading')).to.have.text('Need help?');
-      expect(screen.getByTestId('message')).to.have.text('Ask a staff member.');
-    });
-    it('Renders the footer with custom header props', () => {
-      const fakeStore = {
-        getState: () => ({
-          checkInData: {
-            appointments: [{ clinicPhoneNumber: '555-867-5309' }],
-          },
-        }),
-        subscribe: () => {},
-        dispatch: () => ({}),
-      };
-      const screen = render(
-        <Footer store={fakeStore} header="this is a cool thing" />,
-      );
-      expect(screen.getByTestId('heading')).to.have.text(
-        'this is a cool thing',
-      );
-    });
-    it('Renders the footer with custom message props', () => {
-      const fakeStore = {
-        getState: () => ({
-          checkInData: {
-            appointments: [{ clinicPhoneNumber: '555-867-5309' }],
-          },
-        }),
-        subscribe: () => {},
-        dispatch: () => ({}),
-      };
-      const screen = render(
-        <Footer store={fakeStore} message="this is a cool thing" />,
-      );
-      expect(screen.getByTestId('message')).to.have.text(
-        'this is a cool thing',
-      );
-    });
-    it('Renders the footer when no phone number is presented', () => {
-      const fakeStore = {
-        getState: () => ({
-          checkInData: { appointments: [{ clinicPhoneNumber: '' }] },
-        }),
-        subscribe: () => {},
-        dispatch: () => ({}),
-      };
-      const screen = render(<Footer store={fakeStore} />);
-
-      expect(screen.getByTestId('heading')).to.have.text('Need help?');
-      expect(screen.getByTestId('message')).to.have.text('Ask a staff member.');
-    });
-  });
-});
