+++ src/applications/check-in/day-of/pages/CheckIn/DisplayMultipleAppointments.jsx
@@ -5,7 +5,7 @@ import format from 'date-fns/format';
 import AppointmentListItem from '../../../components/AppointmentDisplay/AppointmentListItem';
 import BackButton from '../../../components/BackButton';
 import BackToHome from '../../../components/BackToHome';
-import Footer from '../../components/Footer';
+import Footer from '../../../components/Footer';
 import { useFormRouting } from '../../../hooks/useFormRouting';
 
 import recordEvent from 'platform/monitoring/record-event';
@@ -68,7 +68,7 @@ const DisplayMultipleAppointments = props => {
           Refresh
         </a>
       </p>
-      <Footer />
+      <Footer isPreCheckIn={false} />
       <BackToHome isPreCheckIn={false} />
     </div>
   );
