+++ src/applications/check-in/day-of/pages/Error.jsx
@@ -2,13 +2,13 @@ import React from 'react';
 
 import ErrorMessage from '../../components/ErrorMessage';
 import BackToHome from '../../components/BackToHome';
-import Footer from '../components/Footer';
+import Footer from '../../components/Footer';
 
 const Error = () => {
   return (
     <div className="vads-l-grid-container vads-u-padding-y--5 ">
       <ErrorMessage />
-      <Footer />
+      <Footer isPreCheckIn={false} />
       <BackToHome isPreCheckIn={false} />
     </div>
   );
