+++ src/applications/check-in/day-of/pages/NextOfKin.jsx
@@ -7,7 +7,7 @@ import { useFormRouting } from '../../hooks/useFormRouting';
 import BackButton from '../../components/BackButton';
 import BackToHome from '../../components/BackToHome';
 import { focusElement } from 'platform/utilities/ui';
-import Footer from '../components/Footer';
+import Footer from '../../components/Footer';
 import { seeStaffMessageUpdated } from '../../actions/day-of';
 import NextOfKinDisplay from '../../components/pages/nextOfKin/NextOfKinDisplay';
 import { makeSelectDemographicData } from '../hooks/selectors';
@@ -67,6 +67,7 @@ const NextOfKin = props => {
           yesAction={yesClick}
           noAction={noClick}
           Footer={Footer}
+          isPreCheckIn={false}
         />
         <BackToHome isPreCheckIn={false} />
       </>
