+++ src/applications/check-in/day-of/pages/SeeStaff.jsx
@@ -2,7 +2,7 @@ import React, { useEffect, useMemo } from 'react';
 import { useSelector } from 'react-redux';
 import PropTypes from 'prop-types';
 import BackToHome from '../../components/BackToHome';
-import Footer from '../components/Footer';
+import Footer from '../../components/Footer';
 import { focusElement } from 'platform/utilities/ui';
 import BackButton from '../../components/BackButton';
 
@@ -29,7 +29,7 @@ const SeeStaff = props => {
         <p>Our staff can help you update your contact information.</p>
       )}
       <TravelPayReimbursementLink />
-      <Footer />
+      <Footer isPreCheckIn={false} />
       <BackToHome isPreCheckIn={false} />
     </div>
   );
