+++ src/applications/check-in/pre-check-in/pages/Demographics/index.jsx
@@ -4,7 +4,7 @@ import { useDispatch, useSelector } from 'react-redux';
 import BackToHome from '../../../components/BackToHome';
 import { useFormRouting } from '../../../hooks/useFormRouting';
 import PropTypes from 'prop-types';
-import Footer from '../../components/Footer';
+import Footer from '../../../components/Footer';
 import BackButton from '../../../components/BackButton';
 import DemographicsDisplay from '../../../components/pages/demographics/DemographicsDisplay';
 import recordEvent from 'platform/monitoring/record-event';
