+++ src/applications/check-in/pre-check-in/pages/EmergencyContact/index.jsx
@@ -9,7 +9,7 @@ import { recordAnswer } from '../../../actions/pre-check-in';
 
 import BackButton from '../../../components/BackButton';
 import BackToHome from '../../../components/BackToHome';
-import Footer from '../../components/Footer';
+import Footer from '../../../components/Footer';
 import EmergencyContactDisplay from '../../../components/pages/emergencyContact/EmergencyContactDisplay';
 
 import { useFormRouting } from '../../../hooks/useFormRouting';
