+++ src/applications/check-in/pre-check-in/pages/Error/index.jsx
@@ -6,7 +6,7 @@ import sub from 'date-fns/sub';
 
 import ErrorMessage from '../../../components/ErrorMessage';
 import BackToHome from '../../../components/BackToHome';
-import Footer from '../../components/Footer';
+import Footer from '../../../components/Footer';
 
 import { makeSelectVeteranData } from '../../../selectors';
