+++ src/applications/check-in/pre-check-in/pages/Validate/index.jsx
@@ -9,7 +9,7 @@ import { createSetSession } from '../../../actions/authentication';
 
 import BackToHome from '../../../components/BackToHome';
 import ValidateDisplay from '../../../components/pages/validate/ValidateDisplay';
-import Footer from '../../components/Footer';
+import Footer from '../../../components/Footer';
 
 import { useFormRouting } from '../../../hooks/useFormRouting';
